+++ lib/phobos_db_checkpoint.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'yaml'
 require 'digest'
 require 'active_record'
@@ -13,10 +15,10 @@ require 'phobos_db_checkpoint/handler'
 require 'phobos_db_checkpoint/actions/retry_failure'
 
 module PhobosDBCheckpoint
-  DEFAULT_DB_DIR = 'db'.freeze
+  DEFAULT_DB_DIR = 'db'
   DEFAULT_MIGRATION_PATH = File.join(DEFAULT_DB_DIR, 'migrate').freeze
-  DEFAULT_DB_CONFIG_PATH = 'config/database.yml'.freeze
-  DEFAULT_POOL_SIZE = 5.freeze
+  DEFAULT_DB_CONFIG_PATH = 'config/database.yml'
+  DEFAULT_POOL_SIZE = 5
 
   class << self
     attr_reader :db_config
@@ -45,14 +47,12 @@ module PhobosDBCheckpoint
 
       @db_config_path ||= ENV['DB_CONFIG'] || DEFAULT_DB_CONFIG_PATH
 
-      configs = YAML.load(ERB.new(File.read(File.expand_path(@db_config_path))).result)
+      configs = YAML.safe_load(ERB.new(File.read(File.expand_path(@db_config_path))).result)
       @db_config = configs[env]
 
       pool_size = @db_config['pool']
 
-      if pool_size.nil? && Phobos.config
-        pool_size = number_of_concurrent_listeners + DEFAULT_POOL_SIZE
-      end
+      pool_size = number_of_concurrent_listeners + DEFAULT_POOL_SIZE if pool_size.nil? && Phobos.config
 
       @db_config.merge!('pool' => pool_size || DEFAULT_POOL_SIZE)
     end
@@ -61,10 +61,12 @@ module PhobosDBCheckpoint
       ActiveRecord::Base.establish_connection(db_config)
     end
 
+    # rubocop:disable Lint/HandleExceptions
     def close_db_connection
       ActiveRecord::Base.connection_pool.disconnect!
     rescue ActiveRecord::ConnectionNotEstablished
     end
+    # rubocop:enable Lint/HandleExceptions
 
     def load_tasks
       @db_dir ||= DEFAULT_DB_DIR