+++ lib/phobos_db_checkpoint/actions/retry_failure.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module PhobosDBCheckpoint
   class RetryFailure
     include PhobosDBCheckpoint::Handler