+++ lib/phobos_db_checkpoint/event.rb
@@ -1,9 +1,11 @@
+# frozen_string_literal: true
+
 module PhobosDBCheckpoint
   class Event < ActiveRecord::Base
     include PhobosDBCheckpoint::EventHelper
     after_initialize :assign_checksum
 
-    scope :order_by_event_time_and_created_at, -> {
+    scope :order_by_event_time_and_created_at, lambda {
       order('event_time desc nulls last', 'created_at desc nulls last')
     }