+++ lib/phobos_db_checkpoint/event_actions.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module PhobosDBCheckpoint
   Ack = Struct.new(:entity_id, :event_time, :event_type, :event_version)
 end