+++ lib/phobos_db_checkpoint/failure.rb
@@ -1,8 +1,10 @@
+# frozen_string_literal: true
+
 module PhobosDBCheckpoint
   class Failure < ActiveRecord::Base
     include PhobosDBCheckpoint::EventHelper
 
-    scope :order_by_event_time_and_created_at, -> {
+    scope :order_by_event_time_and_created_at, lambda {
       order('event_time desc nulls last', 'created_at desc nulls last')
     }