+++ lib/phobos_db_checkpoint/handler.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module PhobosDBCheckpoint
   module Handler
     include Phobos::Handler
@@ -14,11 +16,12 @@ module PhobosDBCheckpoint
       include Phobos::Instrumentation
       include Phobos::Handler::ClassMethods
 
-      def retry_consume?(event, event_metadata, exception)
+      def retry_consume?(_event, event_metadata, _exception)
         return true unless Phobos.config&.db_checkpoint&.max_retries
         event_metadata[:retry_count] < Phobos.config&.db_checkpoint&.max_retries
       end
 
+      # rubocop:disable Style/RedundantBegin
       def around_consume(payload, metadata)
         event = PhobosDBCheckpoint::Event.new(
           topic: metadata[:topic],
@@ -38,12 +41,10 @@ module PhobosDBCheckpoint
           event_action = instrument('db_checkpoint.event_action', event_metadata) do
             begin
               yield
-            rescue => e
-              if retry_consume?(event, event_metadata, e)
-                raise e
-              else
-                Failure.record(event: event, event_metadata: event_metadata, exception: e)
-              end
+            rescue StandardError => e
+              raise e if retry_consume?(event, event_metadata, e)
+
+              Failure.record(event: event, event_metadata: event_metadata, exception: e)
             end
           end
 
@@ -61,6 +62,7 @@ module PhobosDBCheckpoint
         # connections to the pool cached by threads that are no longer alive.
         ActiveRecord::Base.clear_active_connections!
       end
+      # rubocop:enable Style/RedundantBegin
     end
   end
 end