+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## UNRELEASED
+### Changed
+- Introduce rubocop style guide
+
 ## 3.3.0 (2017-10-26)
 
 - [enhancement] Bump Phobos version to 1.5.0 to better support Avro. #34